+++ lib/src/create_report.dart
@@ -909,14 +909,14 @@ Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
     if (pubspec.sdkConstraintStatus.hasOptedIntoNullSafety) {
       if (nullSafetyTags.contains('is:null-safe')) {
         subsection = _Subsection(
-            'Package and dependencies are fully migrated to null-safety!',
+            'Package and dependencies are fully migrated to null safety!',
             explanations.map(_explanationToIssue).toList(),
             maxPoints,
             maxPoints,
             ReportStatus.passed);
       } else {
         subsection = _Subsection(
-            'Package declares support for null-safety, but there are issues.',
+            'Package declares support for null safety, but there are issues.',
             [
               ...explanations.map(_explanationToIssue).toList(),
               // TODO(sigurdm): This is no longer enough, because `pub outdated`
@@ -932,7 +932,7 @@ Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
       }
     } else {
       subsection = _Subsection(
-          'Package does not opt in to null-safety.',
+          'Package does not opt in to null safety.',
           [
             _Issue(
               'Package language version (indicated by the sdk constraint '
@@ -958,7 +958,7 @@ Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
     );
   }
   return _makeSection(
-    title: 'Support sound null-safety',
+    title: 'Support sound null safety',
     maxPoints: maxPoints,
     id: ReportSectionId.nullSafety,
     subsections: [subsection],